Change up what we use compression for.
Instead of trying to compress entire URLs, we'll just focus on domain
names.  So long as we can get the domain name linked to an entity, we
can look up DNS records for addresses, for pages which go into furhter
detail about certificate validation, etc.  DNS TXT records can be much,
much larger, and we only have 40 bytes to work with for certification
transactions so we need to make them count.