Rethink the transaction data piece slightly.
Rather than try to be all things, I've pushed the transaction data model
towards leaning on a more rigid format to make the payload lighter.
We've also trimmed down the domain compression so that we don't handle
absolute URLs or IP addresses or anything.  Everything expects a
hostname since we're kind of predicating our trust model on DNS records.

We now have most of the machinery to generate the specifics of a
transaction, and we've baked in metadata and versioning.  Transaction
data now has a real schema.  Woo!