More revamping and cleanup of transactions.
I added a lot of documentation to existing classes and tried to rename
things thoughtfully.  I also reduced down some existing code, some
inefficient and some simply duplicated.

Hopefully this paints a better picture of the whole transaction concept
being the metadata baked in with the OP_RETURN script, and precisely
what a transaction represents, and why we represent it that way.