Pull in the existing Webpack/React scaffolding we had from the MVP.
This is pulling in the Webpack/React bits we had from an MVP version of
BlockCert that I developed in node.  Due to the nature of the plugin
we're using for Xamarin Studio to support DNX projects, we have to move
the DNX solution a directory below where node_modules gets created.
Otherwise, the mechanism the plugin uses to watch for file changes picks
up allll the little JS file in node_modules, and brings Xamarin to a
screeching halt.  Almost entirely unusable.

This is a temporary solution, hopefully, but isn't too terribly ugly.