Fill out more and more of the BlockCert overview.
As I write this, I'm starting to inherently tweak the protocol slightly
to handle issues that crop up as I'm thinking about how to describe
things.

This will probably be the workflow for a little while: document things,
realize the code as-is is insuffcient, write the fixed version in
documentation first, then update the code to match.  As we get closer to
a solid, 1.0-worthy version of the protocol, this should die down.

We'll also want to figure out a lightweight way to propose changes to
the protocol that can have their own process for examining existing
problems, potential problems, etc, similar to a PEP or JSR.